Add VB support with test. Add default namespace handling. (no support for relative paths yet)
