+++ src/protobuf-net.MSBuild.Test/BuildTests.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Build.Evaluation;
 using Microsoft.Build.Framework;
+using Microsoft.Build.Locator;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -10,7 +11,20 @@ using Xunit.Abstractions;
 
 namespace ProtoBuf.Build
 {
-    public class BuildTests
+
+    class MSBuildFixture : IDisposable
+    {
+        public MSBuildFixture()
+        {
+            MSBuildLocator.RegisterDefaults();
+        }
+
+        void IDisposable.Dispose()
+        {
+        }
+    }
+
+    public class BuildTests : IClassFixture<MSBuildFixture>
     {
         ILogger logger;
         ITestOutputHelper o;
@@ -73,84 +87,15 @@ namespace ProtoBuf.Build
         }
 
         [Fact]
-        public void BuildTest()
+        public void CSharpCodeGenTest()
         {
             var exepath = BuildProject("Data/Proj1/Proj.csproj");
-            Assert.Equal("Hello, World\r\n", GetOutput(exepath, ""));
-        }
-    }
-
-    class XUnitTestLogger : ILogger
-    {
-        static bool DumpEnv = false;
-        static bool Verbose = true;
-
-        ITestOutputHelper o;
-
-        public XUnitTestLogger(ITestOutputHelper o)
-        {
-            this.o = o;
         }
 
-        public LoggerVerbosity Verbosity { get; set; }
-        public string Parameters { get; set; }
-
-        public void Initialize(IEventSource eventSource)
+        [Fact]
+        public void VBCodeGenTest()
         {
-            eventSource.AnyEventRaised += EventSource_AnyEventRaised;
-            eventSource.BuildStarted += EventSource_BuildStarted;
-            eventSource.ProjectStarted += EventSource_ProjectStarted;
-            eventSource.MessageRaised += EventSource_MessageRaised;
-            eventSource.ErrorRaised += EventSource_ErrorRaised;
-        }
-
-        private void EventSource_ErrorRaised(object sender, BuildErrorEventArgs e)
-        {
-            o.WriteLine(e.Message + " " + e.File + "(" + e.LineNumber + "," + e.ColumnNumber + ")");
-        }
-
-        private void EventSource_MessageRaised(object sender, BuildMessageEventArgs e)
-        {
-            if (e.Message.StartsWith("DEBUG:"))
-            {
-                o.WriteLine(e.Message);
-            }
-        }
-
-        private void EventSource_ProjectStarted(object sender, ProjectStartedEventArgs e)
-        {
-            if (DumpEnv)
-            {
-                o.WriteLine("Initial Properties:");
-                foreach (DictionaryEntry kvp in e.Properties)
-                {
-                    o.WriteLine($"{kvp.Key} {kvp.Value}");
-                }
-            }
-        }
-
-        private void EventSource_BuildStarted(object sender, BuildStartedEventArgs e)
-        {
-            if (DumpEnv)
-            {
-                o.WriteLine("Environment:");
-                foreach (var kvp in e.BuildEnvironment)
-                {
-                    o.WriteLine($"{kvp.Key}: {kvp.Value}");
-                }
-            }
-        }
-
-        private void EventSource_AnyEventRaised(object sender, BuildEventArgs e)
-        {
-            if (Verbose)
-            {
-                o.WriteLine(e.Message);
-            }
-        }
-
-        public void Shutdown()
-        {
+            var exepath = BuildProject("Data/Proj2/Proj.vbproj");
         }
     }
 }
