+++ src/protobuf-net.MSBuild.Test/Data/Proj1/Proj.csproj
@@ -1,6 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <RootNamespace>TestNS</RootNamespace>
     <TargetFramework>net461</TargetFramework>
   </PropertyGroup>
