+++ src/protobuf-net.MSBuild.Test/Data/Proj1/Test.proto
@@ -1,4 +1,5 @@
 ﻿syntax = "proto2";
+
 enum Foo {
   A = 1;
   B = 2;
