+++ src/protobuf-net.MSBuild.Test/Data/Proj2/Program.vb
@@ -1,0 +1,8 @@
+﻿Module Program
+    Sub Main(args As String())
+        ' "Proj" namespace component is implied from "Proj.vbproj".
+        Dim x = Proj.Foo.A
+        Dim y = Proj.Bar.E
+
+    End Sub
+End Module
