+++ src/protobuf-net.MSBuild.Test/Data/Proj2/Proj.vbproj
@@ -1,0 +1,8 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net461</TargetFramework>
+  </PropertyGroup>
+
+  <Import Project="../Common.props"/>
+</Project>
