+++ src/protobuf-net.MSBuild/GenerateProtoBufCode.cs
@@ -13,28 +13,45 @@ namespace ProtoBuf.MSBuild
     {
         public ITaskItem[] ProtoDef { get; set; }
 
-
         public ITaskItem[] ImportPaths { get; set; }
 
         public string OutputPath { get; set; }
 
+        public string DefaultNamespace { get; set; }
+
+        public string Language { get; set; }
+
         [Output]
         public ITaskItem[] ProtoCodeFile { get; set; }
 
+        static CommonCodeGenerator GetCodeGenForLanguage(string lang)
+        {
+            if(lang == null)
+                return CSharpCodeGenerator.Default;
+
+            switch (lang)
+            {
+                case "C#":
+                    return CSharpCodeGenerator.Default;
+                case "VB":
+                    return VBCodeGenerator.Default;
+                default:
+                    throw new NotSupportedException("protobuf code generation is not supported for language " + lang);
+            }
+        }
+
         public override bool Execute()
         {
-            Log.LogMessage("Dir: " + Environment.CurrentDirectory);
-            var codegen = CSharpCodeGenerator.Default;
+            var codegen = GetCodeGenForLanguage(Language);
 
             if (ProtoDef == null || ProtoDef.Length == 0)
             {
-                Log.LogMessage("Nothing to process");
                 return true;
-            } 
+            }
 
             var set = new FileDescriptorSet
             {
-                DefaultPackage = "test" // not sure what this is? ns maybe?
+                DefaultPackage = DefaultNamespace
             };
 
             if (ImportPaths == null || ImportPaths.Length == 0)
@@ -67,7 +84,6 @@ namespace ProtoBuf.MSBuild
 
             set.Process();
             var options = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-            //options[lhs.Substring(1)] = rhs;
 
             var codeFiles = new List<ITaskItem>();
             var files = codegen.Generate(set, options: options);
